requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Frontend
streamlit==1.35.0

# LangChain ecosystem
langchain==0.2.1
langchain-community==0.2.1
langchain-groq==0.1.6

# Groq
groq==0.9.0

# Embeddings
sentence-transformers==3.0.0
torch==2.3.0

# Vector database
faiss-cpu==1.8.0

# Document parsing
pypdf==4.2.0
python-docx==1.1.2
markdown==3.6
unstructured==0.14.4

# Data validation
pydantic==2.7.1
pydantic-settings==2.3.0

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
numpy==1.26.4
pandas==2.2.2
tiktoken==0.7.0

# Evaluation
ragas==0.1.9

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7

# Logging
loguru==0.7.2
